refactor(member): 패키지 변경
