+++ src/main/java/com/lhgpds/algometa/exception/GlobalExceptionAdvice.java
@@ -223,6 +223,4 @@ public class GlobalExceptionAdvice {
             httpServletRequest.getRequestURI());
         return ResponseEntity.internalServerError().body(errorResponse);
     }
-
-
 }
