+++ src/main/java/com/lhgpds/algometa/internal/member/domain/vo/Email.java
@@ -1,0 +1,5 @@
+package com.lhgpds.algometa.internal.member.domain.vo;
+
+public class Email {
+
+}
