+++ src/test/groovy/com/lhgpds/algometa/configuration/security/handler/OAuth2SuccessHandlerTest.groovy
@@ -4,10 +4,9 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import com.lhgpds.algometa.controller.auth.dto.TokenDto
 import com.lhgpds.algometa.exception.common.DuplicateException
 import com.lhgpds.algometa.internal.auth.jwt.service.JwtTokenService
-import com.lhgpds.algometa.internal.member.entity.vo.Role
+import com.lhgpds.algometa.internal.member.domain.vo.Role
 import com.lhgpds.algometa.internal.member.service.MemberServiceImpl
 import com.lhgpds.algometa.internal.member.service.dto.MemberDto
-import org.springframework.http.HttpStatus
 import org.springframework.http.MediaType
 import org.springframework.mock.web.MockHttpServletRequest
 import org.springframework.mock.web.MockHttpServletResponse
@@ -17,8 +16,6 @@ import org.springframework.security.core.authority.SimpleGrantedAuthority
 import org.springframework.security.oauth2.core.user.DefaultOAuth2User
 import org.springframework.security.oauth2.core.user.OAuth2User
 import spock.lang.Specification
-
-import javax.servlet.http.Cookie
 
 class OAuth2SuccessHandlerTest extends Specification {
