+++ src/test/java/com/lhgpds/algometa/annotation/WithMockAlgoUser.java
@@ -1,6 +1,6 @@
 package com.lhgpds.algometa.annotation;
 
-import com.lhgpds.algometa.internal.member.entity.vo.Role;
+import com.lhgpds.algometa.internal.member.domain.vo.Role;
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Inherited;
